server/requirements.txt
fastapi>=0.100.0
uvicorn
openai==0.28.1
python-dotenv
requests
passlib[bcrypt]
python-multipart
python-jose[cryptography]
httpx
pytest
sqlalchemy>=2.0.0
pydantic>=2.0.0
aiosqlite
cryptography
celery
redis
prometheus_client
websockets
